Proyecto/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
passlib[bcrypt]
pydantic
PyJWT
